Break into hh and cc
